+++ src/AceCommon.h
@@ -59,6 +59,7 @@ SOFTWARE.
 #include "algorithms/isSorted.h"
 #include "algorithms/linearSearch.h"
 #include "cstrings/copyReplace.h"
+#include "print_utils/printReplaceTo.h"
 
 // Version format: "xx.yy.zz" => xxyyzz (without leading 0)
 #define ACE_COMMON_VERSION 10403